Move NetworkPortsSerializer to non-public-api serialization package
